Adds toolbar actions to action screen
Implements toolbar actions on the action screen based on the PromptResponse.
This change introduces a new `ToolbarAction` data class and updates the
`PromptResponse` model to include a list of `ToolbarAction` objects instead
of a list of strings. This allows for more complex toolbar actions with
associated labels and actions. Also adds the UI for rendering the toolbar and
handling clicks.