Renames app and refactors tenant configuration
Updates the application name to "P4W Mobile" and switches to a new theme "P4WMobile".

Introduces a global constant for the default base URL. The tenant selection screen now uses this constant as a placeholder and default value.

The tenant configuration logic is refactored to use the default base URL if no value is stored in shared preferences.
Tenant name is now nullable to handle cases where the user has not yet configured a tenant.