Replaces LocationWorker with LocationService
Migrates from a WorkManager-based LocationWorker to a foreground LocationService for continuous location updates.

This change improves the reliability and frequency of location updates, which are essential for tracking user activity.
The service runs in the foreground to ensure it is not killed by the system.
A periodic worker is not a great fit for this as it has a minimum interval of 15 minutes and also can be killed by the system at any time
Updates location every minute.

Removes the LocationWorker.