Adds background location tracking
Implements background location tracking using a foreground service with a notification channel.

The location service now creates a notification channel to ensure reliable background execution, especially on newer Android versions. Updates the notification content to provide more informative messages to the user.

Also, the login view model now checks for the old base url if a tenant url is not present, to account for older tenants.