Improves location service reliability
Enhances the location service to improve its reliability and resilience.

Specifically, it adds try-catch blocks to handle potential exceptions during service startup, ensuring that failures are logged and reported to Crashlytics, and prevents the service from crashing.

Also, it ensures the location service is only started when a user is logged in and has location tracking enabled, preventing unnecessary battery consumption and data usage.

It also adds the `FOREGROUND_SERVICE_TYPE_LOCATION` to startForeground call for Android 14+ to comply with new platform requirements.